Update my_solution.rb file with reflection and move the gps6 folder to the correct directory
