+++ week-7/gps6/state_data.rb
